requirements.txt
numpy==1.26.4
torch==2.8.0
nnunetv2==2.5.2
acvl_utils==0.2
opencv-python==4.12.0.88
scikit-learn==1.6.1
trimesh==4.10.0
pyradiomics==3.1.0